+++ delivery/app/schemas/menuSchema.js
@@ -29,6 +29,12 @@ const menuSchema = new mongoose.Schema({
         type: Number,
         unique: false,
         required: true
+    },
+    drink: {
+        type: Boolean,
+        unique: false,
+        required: false,
+        default: false
     }
 });