+++ menu/app.js
@@ -1,10 +1,12 @@
 /**
  * Le fichier principal du microservice des menus.
- * @author GAURE Warren
+ * @author GAURE Warren, AMARA Ahmed
  * @version 1.0
 */
 
 const express = require('express');
+const mongoose = require('mongoose');
+const mongoSanitize = require('express-mongo-sanitize');
 const cors = require('cors');
 
 const swaggerUI = require('swagger-ui-express');
@@ -14,11 +16,21 @@ const swaggerDoc = YAML.load('./swagger.yaml');
 const loggerMiddleware = require('./app/middlewares/loggerMiddleware');
 const authenticationMiddleware = require('./app/middlewares/authenticationMiddleware');
 
+const menuRouter = require('./app/routers/menuRouter');
+
 require("dotenv").config();
 
 const app = express();
 
 const port = process.env.MENU_PORT || 3003;
+
+mongoose.connect(process.env.MONGODB_URI, { useNewUrlParser: true, useUnifiedTopology: true })
+.then(() => {
+    console.log("Connected to MongoDB");
+})
+.catch((error) => {
+    console.error("Error while connecting to MongoDB : ", error);
+});
 
 app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
@@ -30,9 +42,7 @@ app.use('/docs', swaggerUI.serve, swaggerUI.setup(swaggerDoc));
 app.use(loggerMiddleware);
 app.use(authenticationMiddleware);
 
-app.get('/hello', function(req, res) {
-    res.send("Hello World !");
-});
+app.use('/menu', menuRouter);
 
 app.listen(port, function() {
     console.log(`Listens to port ${port}`);