+++ product/app/routers/productRouter.js
@@ -1,6 +1,6 @@
 /**
  * Le routeur pour les articles
- * @author AMARA Ahmed
+ * @author GAURE Warren
  * @version 1.0
 */
 
@@ -10,6 +10,7 @@ const productController = require('../controllers/productController');
 
 router.post('/add', productController.createAndAddProduct);
 router.get('/find', productController.findProduct);
+router.get('/getProducts',productController.getProductsByIds);
 router.get('/metrics', productController.metrics);
 
 module.exports = router;