+++ restaurant/app/services/restaurantService.js
@@ -41,6 +41,10 @@ const findRestaurantByID = async (id) => {
     try {
         const restaurant = await Restaurant.findById(id);
 
+        await Restaurant.populate(restaurant, { path: 'products', model: 'Product' });
+        await Restaurant.populate(restaurant, { path: 'orders', model: 'Order' });
+        await Restaurant.populate(restaurant, { path: 'menus', model: 'Menu' });
+
         return restaurant;
     }
     catch(error) {
@@ -52,7 +56,7 @@ const findRestaurantByID = async (id) => {
  * Fonction permettant de retrouver un restaurant dans la base de données grâce à son nom ou son adresse. 
  * @param {String} name - Le nom du restaurant.
  * @param {String} address - L'adresse du restaurant.
- * @returns {Array} Le restaurant trouvé.
+ * @returns {object} Le restaurant trouvé.
  */
 const findRestaurantByNameOrAddress = async (name, address) => {
     try {
@@ -82,12 +86,14 @@ const createRestaurant = async (name, ownerID, address) => {
         const newRestaurant = new Restaurant({
             name: name,
             ownerID: ownerID,
-            address: address
+            address: address,
         });
 
         await newRestaurant.save();
         // For testing purposes
-        console.log("Restaurant created : " + newRestaurant._id);
+        console.log("Restaurant created : ", newRestaurant._id);
+
+        return newRestaurant;
     }
     catch (error) {
         throw new Error("Error while trying to create a restaurant : " + error.message);
@@ -159,7 +165,7 @@ const addProduct = async (restaurantID, product) => {
             throw new Error("Restaurant not found");
         }
 
-        restaurant.products.push(product);
+        restaurant.products.push(product._id);
 
         await restaurant.save();
     }
@@ -181,7 +187,7 @@ const addOrder = async (restaurantID, order) => {
             throw new Error("Restaurant not found");
         }
 
-        restaurant.orders.push(order);
+        restaurant.orders.push(order._id);
 
         await restaurant.save();
     }
@@ -199,8 +205,11 @@ const addOrder = async (restaurantID, order) => {
 const updateOrderStatus = async (restaurantID, orderID, newStatus) => {
     try {
         const restaurant = await Restaurant.findById(restaurantID);
+        
         const index = restaurant.orders.findIndex((o) => o._id.toString() === orderID.toString());
+        
         restaurant.orders[index].status = newStatus;
+        
         await restaurant.save();
     }
     catch (error) {
@@ -228,8 +237,11 @@ const getOrdersSince = async (restaurantID, numberOfDaysBack) => {
         }
         else {
             const currentDate = new Date();
+            
             const startDate = new Date(currentDate.getTime() - numberOfDaysBack * 24 * 60 * 60 * 1000);
+            
             const orders = restaurant.orders.filter((order) => new Date(order.date).getTime() >= startDate.getTime());
+            
             return orders;
         }
     }
@@ -251,6 +263,22 @@ const getTotalRevenue = async (orders) => {
     }
 
     return totalRevenue;
+};
+
+/**
+ * Fonction permettant d'ajouter un menu au restaurant.
+ * @param {String} restaurantID - L'ID du restaurant.
+ * @param {object} menu - Le menu à ajouter.
+*/
+const addMenu = async (restaurantID, menu) => {
+    try {
+        Restaurant.findByIdAndUpdate(restaurantID, {
+            $addToSet: { menu: menu }
+        });
+    }
+    catch (error) {
+        throw new Error("Error while trying to add a product to a restaurant : " + error.message);
+    }
 };
 
 /**
@@ -304,6 +332,7 @@ module.exports = {
     deleteRestaurant,
     getAllRestaurants,
     addProduct,
+    addMenu,
     addOrder,
     updateOrderStatus,
     getOrdersSince,