+++ users/app/controllers/usersController.js
@@ -6,22 +6,18 @@
 
 const usersService = require('../services/usersService');
 const RESTAURANT_URL = `http://${process.env.RESTAURANT_HOST}:${process.env.RESTAURANT_PORT}/restaurant/`;
-const decodeJWT = require('../utils/decodeToken');
 
 const getUser = async (req, res) => {
-    const accessToken = req.headers.authorization.split(' ')[1];
-    const decodeToken = decodeJWT(accessToken)
-    let userID = decodeToken.id;
-    const userType = decodeToken.type;
+    let userID = req.decoded.id;
+    const userType = req.decoded.type;
     const targetUserID = req.headers["userid"];
 
     try {
-
         if (userType === "TECHNICAL") {
-            throw new Error("Forbidden");
+            throw new Error("Technical trying to get user information");
         }
         else if (userType === "SALES" && (userID === targetUserID)) {
-            throw new Error("Forbidden");
+            throw new Error("Sales trying to get their own information");
         }
         else if (userType === "SALES") {
             userID = targetUserID;
@@ -36,21 +32,23 @@ const getUser = async (req, res) => {
     }
     catch (error) {
         if (error.message === "Invalid user ID") {
-            res.status(400).json({ error: "Invalid user ID" });
+            res.status(400).json({ error: error.message });
         }
         else if (error.message === "User not found") {
-            res.status(404).json({ error: "User not found" });
+            res.status(404).json({ error: error.message });
         }
-        else if (error.message === "Forbidden") {
-            res.status(403).json({ error: "Forbidden" });
+        else if (error.message === "Technical trying to get user information") {
+            res.status(403).json({ error: error.message });
+        }
+        else if (error.message === "Sales trying to get their own information") {
+            res.status(403).json({ error: error.message });
         }
         else {
             console.error("Unexpected error while getting user : ", error);
-            res.status(500).json({ error: "User fetching failed" });
+            return res.status(500).json({ error: "Internal server error" });
         }
     }
 };
-
 
 const getUserByEmail = async (req, res) => {
     if (!req.body) {
@@ -64,7 +62,6 @@ const getUserByEmail = async (req, res) => {
     }
 
     try {
-
         const user = await usersService.getUserByEmail(email);
 
         if (!user) {
@@ -77,28 +74,19 @@ const getUserByEmail = async (req, res) => {
         if (error.message === "User not found") {
             res.status(404).json({ error: "User not found" });
         }
-        else if (error.message === "Forbidden") {
-            res.status(403).json({ error: "Forbidden" });
-        }
         else {
             console.error("Unexpected error while getting user by email : ", error);
-            res.status(500).json({ error: "User fetching failed" });
+            return res.status(500).json({ error: "Internal server error" });
         }
-        
     }
 };
 
 const getAllUsers = async (req, res) => {
-    const accessToken = req.headers.authorization.split(' ')[1];
-    const decodeToken = decodeJWT(accessToken)
-    const userType = decodeToken.type;
+    const userType = req.decoded.type;
 
     try {
         if (userType != "SALES") {
             throw new Error("Invalid user type");
-        }
-        else if (userType === "TECHNICAL") {
-            throw new Error("Forbidden");
         }
 
         const allUsers = await usersService.getAllUsers();
@@ -121,7 +109,7 @@ const getAllUsers = async (req, res) => {
         }
         else {
             console.error("Unexpected error while getting all users : ", error);
-            res.status(500).json({ error: "Users fetching failed" });
+            return res.status(500).json({ error: "Internal server error" });
         }
     }
 };
@@ -130,6 +118,7 @@ const editUser = async (req, res) => {
     if (!req.body) {
         return res.status(400).json({ error: "Required request body is missing" });
     }
+
     const targetUserID = parseInt(req.body["userID"]);
     const firstName = req.body["firstName"];
     const lastName = req.body["lastName"];
@@ -143,10 +132,8 @@ const editUser = async (req, res) => {
     }
 
     const accessToken = req.headers.authorization.split(' ')[1];
-    const decodedToken = decodeJWT(accessToken);
-    const userID = decodedToken.id;
-    const userType = decodedToken.type;
-    let editedUser;
+    const userID = req.decoded.id;
+    const userType = req.decoded.type;
 
     try {
         const userToEdit = await usersService.getUser(targetUserID);
@@ -165,11 +152,33 @@ const editUser = async (req, res) => {
             throw new Error("User trying to edit another user without permission");
         }
         if (userType === "CLIENT" || userType === "DELIVERY" || userType === "DEVELOPER") {
-            editedUser = await usersService.editUser(targetUserID, firstName, lastName, address, email, phoneNumber, encryptedPassword);
+            await usersService.editUser(targetUserID, firstName, lastName, address, email, phoneNumber, encryptedPassword);
         }
         else if (userType === "RESTAURANT") {
-            editedUser = await usersService.editUser(targetUserID, firstName, lastName, address, email, phoneNumber, encryptedPassword);
-            // MEttre la route vers le truc Mongo
+            await usersService.editUser(targetUserID, firstName, lastName, address, email, phoneNumber, encryptedPassword);
+            
+            const restaurantAddress = req.body["restaurantAddress"];
+            const restaurantName = req.body["restaurantName"];
+
+            if (!restaurantAddress || !restaurantName) {
+                throw new Error("Missing mandatory data for restaurant edit");
+            }
+
+            const url = `${RESTAURANT_URL}edit`;
+            const response = await axios.post(url, {
+                restaurantID: targetUserID,
+                address: restaurantAddress,
+                name: restaurantName
+            },
+            { 
+                headers: { 
+                    Authorization: `Bearer ${accessToken}`
+                }
+            });
+
+            if (response.status !== 200) {
+                throw new Error("Restaurant couldn't be edited");
+            }
         }
         
         return res.status(200).json({ message: "User edited" });
@@ -188,14 +197,20 @@ const editUser = async (req, res) => {
             res.status(403).json({ error: "Forbidden" });
         }
         else if (error.message === "User trying to change their own type") {
-            res.status(403).json({ error: "Forbidden" });
+            res.status(403).json({ error: error.message });
         }
         else if (error.message === "User trying to change their own suspension status") {
-            res.status(403).json({ error: "Forbidden" });
+            res.status(403).json({ error: error.message });
+        }
+        else if (error.message === "Missing mandatory data for restaurant edit") {
+            res.status(400).json({ error: error.message });
+        }
+        else if (error.message === "Restaurant couldn't be edited") {
+            res.status(500).json({ error: error.message });
         }
         else {
             console.error("Unexpected error while editing user : ", error);
-            res.status(500).json({ error: "User editing failed" });
+            return res.status(500).json({ error: "Internal server error" });
         }
     }
 };
@@ -211,8 +226,7 @@ const suspendUser = async (req, res) => {
         return res.status(400).json({ error: "Missing mandatory data for suspension" });
     }
 
-    const accessToken = req.headers.authorization.split(' ')[1];
-    const userType = decodeJWT(accessToken).type;
+    const userType = req.decoded.type;
 
     try {
         if (userType != "SALES") {
@@ -228,7 +242,7 @@ const suspendUser = async (req, res) => {
             throw new Error("Wrong user ID in request body");
         }
         else if (userToSuspend.userType === "SALES") {
-            throw new Error("User trying to suspend another commercial service user");
+            throw new Error("User trying to suspend another sales service user");
         }
         else if (userToSuspend.userType === "TECHNICAL") {
             throw new Error("User trying to suspend a technical service user");
@@ -242,24 +256,21 @@ const suspendUser = async (req, res) => {
         if (error.message === "Invalid user type") {
             res.status(403).json({ error: "Forbidden" });
         }
-        else if (error.message === "Forbidden") {
-            res.status(403).json({ error: "Forbidden" });
-        }
         else if (error.message === "User not found") {
             res.status(404).json({ error: "User not found" });
         }
         else if (error.message === "Wrong user ID in request body") {
             res.status(400).json({ error: "Wrong user ID in request body" });
         }
-        else if (error.message === "User trying to suspend another commercial service user") {
-            res.status(403).json({ error: "Forbidden" });
+        else if (error.message === "User trying to suspend another sales service user") {
+            res.status(403).json({ error: error.message });
         }
         else if (error.message === "User trying to suspend a technical service user") {
-            res.status(403).json({ error: "Forbidden" });
+            res.status(403).json({ error: error.message });
         }
         else {
             console.error("Unexpected error while suspending user : ", error);
-            res.status(500).json({ error: "User suspension failed" });
+            return res.status(500).json({ error: "Internal server error" });
         }
     }
 };
@@ -275,8 +286,7 @@ const unsuspendUser = async (req, res) => {
         return res.status(400).json({ error: "Missing mandatory data for suspension" });
     }
 
-    const accessToken = req.headers.authorization.split(' ')[1];
-    const userType = decodeJWT(accessToken).type;
+    const userType = req.decoded.type;
 
     try {
         if (userType != "SALES") {
@@ -302,7 +312,7 @@ const unsuspendUser = async (req, res) => {
         }
         else {
             console.error("Unexpected error while suspending user : ", error);
-            res.status(500).json({ error: "User suspension failed" });
+            return res.status(500).json({ error: "Internal server error" });
         }
     }
 };
@@ -319,14 +329,14 @@ const deleteUser = async (req, res) => {
     }
 
     const accessToken = req.headers.authorization.split(' ')[1];
-    const decodedToken = decodeJWT(accessToken);
-    const userID = decodedToken.id;
-    const userType = decodedToken.type;
+    const userID = req.decoded.id;
+    const userType = req.decoded.type;
 
     try {
         if (userType === "TECHNICAL") {
             throw new Error("Invalid user type");
         }
+
         const targetUser = await usersService.getUser(targetUserID);
 
         if (!targetUser) {
@@ -347,14 +357,13 @@ const deleteUser = async (req, res) => {
 
         await usersService.deleteUser(targetUserID);
 
-        // Delete le restaurant si delete le restaurateur
         if (targetUser.userType === "RESTAURANT") {
-            // appelle a la route
-            url = `${RESTAURANT_URL}find`;
-            response = await axios.get(url, {
+            url = `${RESTAURANT_URL}delete`;
+            response = await axios.delete(url, {
                 params: { id: targetUserID },
                 headers: { Authorization: `Bearer ${accessToken}` }
             });
+
             if (response.status !== 200) {
                 throw new Error("Restaurant couldn't be deleted");
             }
@@ -364,32 +373,35 @@ const deleteUser = async (req, res) => {
     }
     catch (error) {
         if (error.message === "User not found") {
-            res.status(404).json({ error: "User not found" });
+            res.status(404).json({ error: error.message });
         }
         else if (error.message === "Wrong user ID in request body") {
-            res.status(400).json({ error: "Wrong user ID in request body" });
+            res.status(400).json({ error: error.message });
         }
         else if (error.message === "User trying to delete another user without permission") {
-            res.status(403).json({ error: "Forbidden" });
+            res.status(403).json({ error: error.message });
         }
         else if (error.message === "Commercial trying to delete themselves") {
-            res.status(403).json({ error: "Forbidden" });
+            res.status(403).json({ error: error.message });
         }
         else if (error.message === "Technical trying to delete themselves") {
-            res.status(403).json({ error: "Forbidden" });
+            res.status(403).json({ error: error.message });
         }
         else if (error.message === "Invalid user type") {
+            res.status(403).json({ error: error.message });
+        }
+        else if (error.message === "Restaurant couldn't be deleted") {
+            res.status(500).json({ error: error.message });
         }
         else {
             console.error("Unexpected error while editing user : ", error);
-            res.status(500).json({ error: "User editing failed" });
+            return res.status(500).json({ error: "Internal server error" });
         }
     }
 };
 
 const metrics = async (req, res) => {
-    const accessToken = req.headers.authorization.split(' ')[1];
-    const userType = decodeJWT(accessToken).type;
+    const userType = req.decoded.type;
 
     try {
         if (userType != "TECHNICAL") {
@@ -406,7 +418,7 @@ const metrics = async (req, res) => {
         }
         else {
             console.error("Unexpected error while getting metrics : ", error);
-            res.status(500).json({ error: "Metrics collecting failed" });
+            return res.status(500).json({ error: "Internal server error" });
         }
     }
 };